requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.18.5
pickle-mixin==1.0.2
pkg-resources==0.0.0
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
six==1.15.0
